added useFormValidator
